% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.acmpca_operations.R
\name{create_certificate_authority}
\alias{create_certificate_authority}
\title{Creates a private subordinate certificate authority (CA)}
\usage{
create_certificate_authority(CertificateAuthorityConfiguration,
  RevocationConfiguration = NULL, CertificateAuthorityType,
  IdempotencyToken = NULL, Tags = NULL)
}
\arguments{
\item{CertificateAuthorityConfiguration}{[required] Name and bit size of the private key algorithm, the name of the signing algorithm, and X.500 certificate subject information.}

\item{RevocationConfiguration}{Contains a Boolean value that you can use to enable a certification revocation list (CRL) for the CA, the name of the S3 bucket to which ACM PCA will write the CRL, and an optional CNAME alias that you can use to hide the name of your bucket in the \strong{CRL Distribution Points} extension of your CA certificate. For more information, see the CrlConfiguration structure.}

\item{CertificateAuthorityType}{[required] The type of the certificate authority. Currently, this must be \strong{SUBORDINATE}.}

\item{IdempotencyToken}{Alphanumeric string that can be used to distinguish between calls to \strong{CreateCertificateAuthority}. Idempotency tokens time out after five minutes. Therefore, if you call \strong{CreateCertificateAuthority} multiple times with the same idempotency token within a five minute period, ACM PCA recognizes that you are requesting only one certificate. As a result, ACM PCA issues only one. If you change the idempotency token for each call, however, ACM PCA recognizes that you are requesting multiple certificates.}

\item{Tags}{Key-value pairs that will be attached to the new private CA. You can associate up to 50 tags with a private CA.}
}
\description{
Creates a private subordinate certificate authority (CA). You must specify the CA configuration, the revocation configuration, the CA type, and an optional idempotency token. The CA configuration specifies the name of the algorithm and key size to be used to create the CA private key, the type of signing algorithm that the CA uses to sign, and X.500 subject information. The CRL (certificate revocation list) configuration specifies the CRL expiration period in days (the validity period of the CRL), the Amazon S3 bucket that will contain the CRL, and a CNAME alias for the S3 bucket that is included in certificates issued by the CA. If successful, this operation returns the Amazon Resource Name (ARN) of the CA.
}
\section{Accepted Parameters}{
\preformatted{create_certificate_authority(
  CertificateAuthorityConfiguration = list(
    KeyAlgorithm = "RSA_2048"|"RSA_4096"|"EC_prime256v1"|"EC_secp384r1",
    SigningAlgorithm = "SHA256WITHECDSA"|"SHA384WITHECDSA"|"SHA512WITHECDSA"|"SHA256WITHRSA"|"SHA384WITHRSA"|"SHA512WITHRSA",
    Subject = list(
      Country = "string",
      Organization = "string",
      OrganizationalUnit = "string",
      DistinguishedNameQualifier = "string",
      State = "string",
      CommonName = "string",
      SerialNumber = "string",
      Locality = "string",
      Title = "string",
      Surname = "string",
      GivenName = "string",
      Initials = "string",
      Pseudonym = "string",
      GenerationQualifier = "string"
    )
  ),
  RevocationConfiguration = list(
    CrlConfiguration = list(
      Enabled = TRUE|FALSE,
      ExpirationInDays = 123,
      CustomCname = "string",
      S3BucketName = "string"
    )
  ),
  CertificateAuthorityType = "SUBORDINATE",
  IdempotencyToken = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

