% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.acmpca_operations.R
\name{create_certificate_authority_audit_report}
\alias{create_certificate_authority_audit_report}
\title{Creates an audit report that lists every time that the your CA private key is used}
\usage{
create_certificate_authority_audit_report(
  CertificateAuthorityArn = "string",
  S3BucketName = "string",
  AuditReportResponseFormat = "JSON"|"CSV"
)
}
\arguments{
\item{CertificateAuthorityArn}{[required] Amazon Resource Name (ARN) of the CA to be audited. This is of the form:

\code{arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 region:account:certificate-authority/12345678-1234-1234-1234-123456789012 }.}

\item{S3BucketName}{[required] Name of the S3 bucket that will contain the audit report.}

\item{AuditReportResponseFormat}{[required] Format in which to create the report. This can be either \strong{JSON} or \strong{CSV}.}
}
\description{
Creates an audit report that lists every time that the your CA private key is used. The report is saved in the Amazon S3 bucket that you specify on input. The IssueCertificate and RevokeCertificate operations use the private key. You can generate a new report every 30 minutes.
}
\examples{

}
