% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.acmpca_operations.R
\name{describe_certificate_authority_audit_report}
\alias{describe_certificate_authority_audit_report}
\title{Lists information about a specific audit report created by calling the CreateCertificateAuthorityAuditReport operation}
\usage{
describe_certificate_authority_audit_report(CertificateAuthorityArn,
  AuditReportId)
}
\arguments{
\item{CertificateAuthorityArn}{[required] The Amazon Resource Name (ARN) of the private CA. This must be of the form:

\code{arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 region:account:certificate-authority/12345678-1234-1234-1234-123456789012 }.}

\item{AuditReportId}{[required] The report ID returned by calling the CreateCertificateAuthorityAuditReport operation.}
}
\description{
Lists information about a specific audit report created by calling the CreateCertificateAuthorityAuditReport operation. Audit information is created every time the certificate authority (CA) private key is used. The private key is used when you call the IssueCertificate operation or the RevokeCertificate operation.
}
\section{Accepted Parameters}{
\preformatted{describe_certificate_authority_audit_report(
  CertificateAuthorityArn = "string",
  AuditReportId = "string"
)
}
}

