% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.acmpca_operations.R
\name{get_certificate}
\alias{get_certificate}
\title{Retrieves a certificate from your private CA}
\usage{
get_certificate(
  CertificateAuthorityArn = "string",
  CertificateArn = "string"
)
}
\arguments{
\item{CertificateAuthorityArn}{[required] The Amazon Resource Name (ARN) that was returned when you called CreateCertificateAuthority. This must be of the form:

\code{arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 region:account:certificate-authority/12345678-1234-1234-1234-123456789012 }.}

\item{CertificateArn}{[required] The ARN of the issued certificate. The ARN contains the certificate serial number and must be in the following form:

\code{arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012/certificate/286535153982981100925020015808220737245 region:account:certificate-authority/12345678-1234-1234-1234-123456789012/certificate/286535153982981100925020015808220737245 }}
}
\description{
Retrieves a certificate from your private CA. The ARN of the certificate is returned when you call the IssueCertificate operation. You must specify both the ARN of your private CA and the ARN of the issued certificate when calling the \strong{GetCertificate} operation. You can retrieve the certificate if it is in the \strong{ISSUED} state. You can call the CreateCertificateAuthorityAuditReport operation to create a report that contains information about all of the certificates issued and revoked by your private CA.
}
\examples{

}
