% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.acmpca_operations.R
\name{list_certificate_authorities}
\alias{list_certificate_authorities}
\title{Lists the private certificate authorities that you created by using the CreateCertificateAuthority operation}
\usage{
list_certificate_authorities(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{Use this parameter when paginating results in a subsequent request after you receive a response with truncated results. Set it to the value of the \code{NextToken} parameter from the response you just received.}

\item{MaxResults}{Use this parameter when paginating results to specify the maximum number of items to return in the response on each page. If additional items exist beyond the number you specify, the \code{NextToken} element is sent in the response. Use this \code{NextToken} value in a subsequent request to retrieve additional items.}
}
\description{
Lists the private certificate authorities that you created by using the CreateCertificateAuthority operation.
}
\examples{

}
