% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.acmpca_operations.R
\name{list_tags}
\alias{list_tags}
\title{Lists the tags, if any, that are associated with your private CA}
\usage{
list_tags(CertificateAuthorityArn, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{CertificateAuthorityArn}{[required] The Amazon Resource Name (ARN) that was returned when you called the CreateCertificateAuthority operation. This must be of the form:

\code{arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 region:account:certificate-authority/12345678-1234-1234-1234-123456789012 }}

\item{NextToken}{Use this parameter when paginating results in a subsequent request after you receive a response with truncated results. Set it to the value of \strong{NextToken} from the response you just received.}

\item{MaxResults}{Use this parameter when paginating results to specify the maximum number of items to return in the response. If additional items exist beyond the number you specify, the \strong{NextToken} element is sent in the response. Use this \strong{NextToken} value in a subsequent request to retrieve additional items.}
}
\description{
Lists the tags, if any, that are associated with your private CA. Tags are labels that you can use to identify and organize your CAs. Each tag consists of a key and an optional value. Call the TagCertificateAuthority operation to add one or more tags to your CA. Call the UntagCertificateAuthority operation to remove tags.
}
\section{Accepted Parameters}{
\preformatted{list_tags(
  CertificateAuthorityArn = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

