% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.acmpca_operations.R
\name{update_certificate_authority}
\alias{update_certificate_authority}
\title{Updates the status or configuration of a private certificate authority (CA)}
\usage{
update_certificate_authority(CertificateAuthorityArn,
  RevocationConfiguration = NULL, Status = NULL)
}
\arguments{
\item{CertificateAuthorityArn}{Amazon Resource Name (ARN) of the private CA that issued the certificate to be revoked. This must be of the form:

\code{arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 region:account:certificate-authority/12345678-1234-1234-1234-123456789012 }}

\item{RevocationConfiguration}{Revocation information for your private CA.}

\item{Status}{Status of your private CA.}
}
\description{
Updates the status or configuration of a private certificate authority (CA). Your private CA must be in the \code{ACTIVE} or \code{DISABLED} state before you can update it. You can disable a private CA that is in the \code{ACTIVE} state or make a CA that is in the \code{DISABLED} state active again.
}
\examples{

}
